requirements.txt
allure-pytest==2.13.2
allure-python-commons==2.13.2
pytest==7.4.0
pytest-cov==4.1.0
python-dateutil==2.8.2
requests==2.31.0